자유게시판 - 드래그앤드랍 첨부파일 1.게시글등록 - 첨부파일 구현 완료 2. 게시글 답글 - 첨부파일 구현완료 3. 상세게시글 - 첨부파일 구현완료
게시글 삭제
1.게시글 show = n 으로 변경
2.로컬에 저장된 첨부파일 삭제(Ajax와 Array사용)
3.DB에 저장된 첨부파일 삭제

자유게시판 - 드래그앤드롭 첨부파일
 - 게시글 수정
  1.View.jsp
   : 사용자가 첨부파이 X버튼을 클릭하면 -> 디자인만 삭제. 삭제목록에 저장
   : 사용자가 드래그앤드롭으로 새로운 첨부파일 올리면 -> Local 저장
  2.수정버튼 클릭
   : Ajax로 삭제목록에 있는 첨부파일을 Local에서 삭제
  3. 서버
   : DB에서 게시글 수정 + filecnt도 같이 수정
   : 무조건 해당 게시글의 첨부파일 목록 DB에서 삭제
   : 수정버튼 클릭시 서버단으로 전송된 첨부파일이 없으면 여기서 종료
   : 수정버튼 킬릭시 서버단으로 전송된 첨부파일이 있으면 tbl_attach에 새로 등록

+++ yellow/src/main/webapp/WEB-INF/views/board/write.jsp
@@ -151,7 +151,7 @@
 						<td>글 작성</td>
 						<td>
 							<script type="text/javascript" src="${path}/resources/smarteditor/js/service/HuskyEZCreator.js" charset="utf-8"></script>
-							<textarea cols="93" rows="28" placeholder="게시글을 입력하세요" id="board_content" name="view_content" style="min-width:650px;"></textarea>
+							<textarea cols="93" rows="28" placeholder="게시글을 입력하세요" id="board_content" name="view_content" style="min-width:650px;">${one.view_content}</textarea>
 							
 						</td>
 					</tr>
@@ -198,8 +198,7 @@
 				<i class="fas fa-times"></i>
 			</span>
 		</div>
-	</li>
-		
+	</li>		
 </script>
 <script src="${path}/resources/js/fileAttach.js"></script>
 
@@ -208,6 +207,10 @@
 	console.log('flag:' + flag);
 	//Handlebars 파일 템플릿 컴파일
 	var fileTemplate = Handlebars.compile($("#fileTemplate").html());
+	
+	//수정시 로컬에서 삭제할 기존 천부파일 목록
+	var deleteFileList = new Array();
+	
 	$(function(){
 	//	alert('데이터: ' +${one}); // 공백으로 뜬다는거는 그 코드를 그대로 써도 된다는 것 
 		//register ==> 게시글 등록과 게시글 수정
@@ -221,6 +224,7 @@
 			$('.board_type').val('${one.type}').attr('selected','selected');// 셀렉트박스에서 그녀석을 셀렉트해라
 							//free인것을 셀렉트해주세요
 			$('.up_time')	
+			listAttach('${path}','${one.bno}');
 		}else if(flag == 'answer'){
 			$('.content_header').text('게시글 답글');// 글씨 바꾸는거 .text를 이용해서
 			$('.up_btn').text('답글');
@@ -279,10 +283,15 @@
 					}, error:function(){
 						alert('system error');
 					}
-				});
-				
+				});				
 			}else{// 게시글 수정
-									    
+				var arr_size = deleteFileList.length;				
+				deleteFileList[arr_size] = $(this).attr('data-src');			
+				$(this).parents('li').next('input').remove();
+				$(this).parents('li').remove();
+				
+				for(var i = 0; i < deleteFileList.length; i++)
+					console.log(i+',' + deleteFileList[i]);
 			}
 		});
 					
@@ -348,9 +357,11 @@
 			});
 			//로컬 드라이브에 저장되어 있는 해당 게시글
 			//첨부파일 삭제
-/* 			if(deletedFile.length > 0){
+			
+			//삭제 첨부파일 목록에 있는 첨부파일들 local에서 삭제
+			if(deleteFileList.length > 0){
 				$.post('${path}/upload/deleteAllFile', {files:deleteFileList},function(){});
-			} */
+			} 
 			//폼에 hidden 태그들을 붙임
 			$("#frm_board").append(str); //
